Created new Rollercoaster subclasses. Toggle now decides which rollercoaster to draw
