+++ src/js/constants/rollerCoasterTypes.js
@@ -1,17 +1,18 @@
 export default {
     'No Cruzado' : [
-        [0, 0, 0], [0, 0, 5],
-            [5, 0, 5], [5, 0, 0], [5, 0, -5], [5, 10, -10],
-            [7, 10, -20], [5, 5, -20], [5, 6, -30], [5, 0, -35],
-            [5, 0, -40], [-10, 0, -40], [-10, 2, -35], [-10, 0, -30],
-            [0, 0, -30], [0, 0, -20], [-10, 3, -20], [-10, 3, -10], 
-            [-10, 0, -5], [-5, 0, -5], [-5, 0, 0],
-            [0, 0, 0], [0, 0, 5], [5, 0, 5],
+        [0, 0, 0], [0, 0, 5], [5, 0, 5], [5, 0, 0],
+        [5, 0, -5], [5, 10, -10], [7, 10, -20], [5, 5, -20], 
+        [5, 6, -30], [5, 0, -35], [5, 0, -40], [-10, 0, -40],
+        [-10, 2, -35], [-10, 0, -30], [0, 0, -30], [0, 0, -20], 
+        [-10, 3, -20], [-10, 3, -10], [-10, 0, -5], [-5, 0, -5], 
+        [-5, 0, 0], [0, 0, 0], [0, 0, 5], [5, 0, 5],
     ],
 
     'Cruzado' : [
-        [0, 0, 0], [0, 0, 1], [0, 0, 3], [0, 0, 5],
-        [5, 0, 5], [5, 0, 0], [5, 0, -5], [5, 10, -10],
-        
+        [0, 0, 0], [0, 0, 5], [10, 0, 5], [10, 0, 0],
+        [7.5, 5, -5], [5, 10, -10], [2.5, 15, -15], [0, 5, -30], 
+        [-5, 5, -30], [-10, 0, -40], [10, 0, -40], [10, 7, -20], 
+        [0, 0, -20], [-5, 0, -20], [-10, 3, -10], [-5, 5, -5], 
+        [-5, 0, 0], [0, 0, 0], [0, 0, 5], [10, 0, 5]
     ]
 }
