+++ src/js/models/scene.js
@@ -1,7 +1,8 @@
 import TreeNode from './tree_node'
 import Floor from './floor'
 import Skybox from './skybox'
-import Rollercoaster from './rollercoaster'
+import CircularRollercoaster from './circular_rollercoaster'
+import CrossedRollercoaster from './crossed_rollercoaster'
 import { mat4 } from 'gl-matrix'
 import Carousel from './carousel'
 
@@ -10,7 +11,8 @@ export default class Scene extends TreeNode {
         super();
         this.floor = new Floor();
         this.skybox = new Skybox();
-        this.coaster = new Rollercoaster();
+        this.circularCoaster = new CircularRollercoaster();
+        this.crossedCoaster = new CrossedRollercoaster();
         this.carousel = new Carousel();
     }
 
@@ -20,7 +22,8 @@ export default class Scene extends TreeNode {
 
         let coasterModelMatrix = mat4.clone(modelMatrix);
         mat4.translate(coasterModelMatrix, coasterModelMatrix, [15, 0, -10]);
-        this.coaster.draw(coasterModelMatrix, viewMatrix, projMatrix);
+        if (window.tipoDeCamino === "No Cruzado") this.circularCoaster.draw(coasterModelMatrix, viewMatrix, projMatrix);
+        if (window.tipoDeCamino === "Cruzado") this.crossedCoaster.draw(coasterModelMatrix, viewMatrix, projMatrix);
 
         let carouselModelMatrix = mat4.clone(modelMatrix);
         mat4.translate(carouselModelMatrix, carouselModelMatrix, [-3, 0, -10]);
